pip-requirements.txt
asgiref
Django
pip
psycopg2
pytz
setuptools
sqlparse
djangorestframework
request